Update re: New Context AWS actuator.
Per John-Mark Gurney email to OpenC2 IC-SC:  Thanks to support from Duncan Sparrell of sFractal Consulting, New Context has updated the OpenC2 AWS Actuator to support OpenC2 v1.0.  This is done by updating it to use the latest lycan library.  It also now works on Python 3.6 (and likely later versions), and follows most of the HTTPS specification.  There are a few minor things missing, such as TLS support and authentication, but neither of those are difficult to add.